+++ apps/sim/providers/utils.ts
@@ -51,7 +51,7 @@ import {
 } from '@/providers/models'
 import type { ProviderId, ProviderToolConfig } from '@/providers/types'
 import { useProvidersStore } from '@/stores/providers/store'
-import { mergeToolParameters } from '@/tools/params'
+import { mergeToolParameters } from '@/tools/merge-params'
 
 const logger = createLogger('ProviderUtils')
